+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- **\[Python/C++\]** Add `point_format_id` argument to `Writer` constructor to allow update of the point format between reader and writer.
+
 ## [2.1.3] - 2021-11-17
 
 ### Changed